Window focus swaps on click finished(?). A bit janky, so I included full comment at the start of Notepad.tsx file.
Basically, base z-index is
.window -> .notepad -> maximized .window -> .active-window

Clicking any window sets it as .active-window (z-index: 903), this is good however problematic when clicking a notepad file - this sets the file explorer/documents window as .active-window, which means the then rendering notepad rendering below.

As a fix, when a notepad instance renders we just clear .active-window from every .window element, that way notepad's explicit z-index 901 takes over and renders on top of the file explorer. It will also render on top of older notepad windows due to hierarchy.

I also accounted for the fact that I might allow focusing windows on top of maximized windows later on, so .active-window renders on top of .maximized windows.